ComptonCamera6/exporter.py: bugfixes + eventnumber
Read GlobalEventNumber from simulation file instead of just using
increasing numbers.

Save classification result in ClassID instead of EventType

BUGFIXES:
1. rotate vertex position into Luebeck coordinate system
2. have correct number of arry elements if no simulation file is given

TODO:
Something seems strange with ClassID, as it doesn't seem to match with
other variables. For example I would expect the vertex to be non-zero
for good Compton events but this is not always the case.