ComptonCamera6/exporter.py: array of selected events
Added the possibility to pass on the array of selected events which
states which events are classified as Compton by the NN.
This is needed additionally if values from the simulation file should
be stored in order to select the correct events from the root tree.
Before, inconsistent files where produced in which the events where not
machted properly.