ComptonCamera6/exporter: bugfix coordinate transformation
The coordinate transformations into the Luebeck coordinate
system where wrong. They assumed the data to be in the
cracow system while they are now provided in the Aachen
coordinate system. This is fixed.

Attentions has to be paid in case simulation files are also
processed: the old 1-to-1 coupling files still use the Cracow
system instead of the Aachen one. To process the new 4-to-1
simulation files, further changes are needed.